aws_pdf_rag_quickstart/pyproject.toml
[tool.poetry]
name = "aws-rag-quickstart"
version = "0.1.0"
description = "Prototype for AWS RAG Quickstart"
authors = ["Chris King <cking42976@gmail.com>"]
readme = "README.md"
packages = [{include = "aws_rag_quickstart", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
chainlit = "^0.7.0"
transformers = "^4.36.0"
torch = "^2.1.0"

[project]
name = "aws-rag-quickstart"
version = "0.1.0"
description = "Prototype for AWS RAG Quickstart"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.4",
    "pdf2image==1.17.0",
    "boto3==1.34.162",
    "langchain-aws>=0.1.0",
    "opensearch-py~=2.7.1",
    "langchain-community>=0.2.17",
    "requests-aws4auth~=1.3.1",
    "python-dotenv~=1.0.1",
    "langchain-ollama~=0.2.0",
    "ollama~=0.3.3",
    "pypdf2>=3.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-mock==3.14.0",
    "pylama==8.4.1",
    "black==24.10.0",
    "isort==5.13.2",
    "setuptools",
    "coverage",
    "pre-commit",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
